improve error handling for dependency initializing add dependency initializer spec move related ctes into subfolder
