--- app/modules/report/ctes/ts_range_and_interval.rb
@@ -1,25 +1,0 @@
-# frozen_string_literal: true
-
-module Report
-  module Ctes
-    class TsRangeAndInterval < Report::Cte::NodeTemplate
-
-      table_name :time_range_and_interval
-
-      options do
-        {
-          start_time: Time.utc(2000, 1, 1),
-          end_time: Time.utc(2000, 1, 7),
-          interval: '1 day'
-        }
-      end
-
-      select do
-        lower, upper, interval = options.values_at(:start_time, :end_time, :interval)
-        arel_tsrange_expr = Report::TimeSeries.arel_tsrange(lower, upper).as('time_range')
-        bucket_interval_expr = Report::TimeSeries.arel_interval(interval).as('bucket_interval')
-        Arel::SelectManager.new.project(arel_tsrange_expr, bucket_interval_expr)
-      end
-    end
-  end
-end
