+++ spec/lib/modules/report/audio_event_summary_spec.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-describe 'Audio Events Report' do
+describe 'Audio Event Summary' do
   include SqlHelpers::Example
-  subject { Report::AudioEvents.new(options: params) }
+  subject { Report::Ctes::AudioEventSummary.new(options: params) }
 
   before do
     user = create(:user)
@@ -14,13 +14,14 @@ describe 'Audio Events Report' do
       start_time: Time.new('2000-03-26T07:06:59').iso8601,
       end_time: Time.new('2000-04-26T07:06:59').iso8601,
       scaling_factor: 1920,
-      bucket_size: 'day'
+      lower_field: :recorded_date,
+      upper_field: :end_date,
+      interval: '1 day'
     }
   }
 
-  it 'does something' do
-    debugger
-    expect(subject).to be_a(Report::AudioEvents)
-    subject.to_sql
+  it 'executes' do
+    expect(subject).to be_a(Report::Ctes::AudioEventSummary)
+    expect(subject.execute).to be_a(PG::Result)
   end
 end
